requirements.txt
flask
python-dotenv
SQLAlchemy
PyMySQL
flask-migrate
flask-bcrypt
flask_jwt_extended
